requirements.txt
langchain
langchain-community
langchain-google-genai
ollama
gradio
numpy
python-dotenv